Revert "test(backup): Create first model test (#52542)"
This reverts commit 8a99a9b7394faffbb9c5ef50bc27a574fcbcaff2.
